% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airbnb_tokens}
\alias{airbnb_tokens}
\title{Dutch reviews of AirBnB customers on Brussels address locations tokenised using udpipe}
\description{
This data.frame was obtained by using the udpipe R package in order
to obtain tokens, lemmas, part of speech tags and dependency relations of the text in the \code{\link{airbnb}} dataset.
The dataset contains one row per document/token and in addition shows the start and end position of where the token was
found in the original text as well as all information udpipe returns when doing the annotation.\cr
This data is distributed under the CC BY-SA 4.0 license \url{https://creativecommons.org/licenses/by-sa/4.0} 
as the udpipe model which was used to construct the data also has that license.
To see how the data was generated, look at the code in the examples.
\itemize{
 \item{doc_id: }{a unique identifier of the review, which is also available in \code{\link{airbnb}}}
 \item{start: }{the starting position in \code{text} of \code{\link{airbnb}} where the \code{token} is found}
 \item{end: }{the end position in \code{text} of \code{\link{airbnb}} where the \code{token} is found}
 \item{token: }{the token}
 \item{...: }{other udpipe fields like lemma/upos/xpos/dependency relations, see the help of the udpipe R package for details on the output}
}
}
\examples{
data(airbnb_tokens)
str(airbnb_tokens)
head(airbnb_tokens)

\dontrun{
##
## This dataset was obtained as follows using udpipe R package version >= 0.7 
##
library(udpipe)
udmodel <- udpipe_download_model("dutch", 
                                 udpipe_model_repo = "bnosac/udpipe.models.ud")
udmodel <- udpipe_load_model(udmodel$file_model)
x <- udpipe_annotate(udmodel, x = airbnb$text, doc_id = airbnb$doc_id)
x <- as.data.frame(x, detailed = TRUE)
}
}
\references{
\url{https://github.com/UniversalDependencies/UD_Dutch-Alpino/blob/master/LICENSE.txt}
}
\seealso{
\code{\link{airbnb_tokens}}
}
