% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{as.crf}
\alias{as.crf}
\title{Convert a model built with CRFsuite to an object of class crf}
\usage{
as.crf(file, ...)
}
\arguments{
\item{file}{the path to a file on disk containing the CRFsuite model}

\item{...}{other arguments which can be set except the path to the file, namely method, type, options, attribute_names, log (expert use only)}
}
\value{
an object of class \code{crf}
}
\description{
If you have a model built with CRFsuite either by this R package
or by another software library which wraps CRFsuite (e.g. Python/Java), you can 
convert it to an object of class \code{crf} which this package can use 
to inspect the model and to use it for prediction (if you can mimic the way the attributes are created).\cr
This is for expert use only.
}
