% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{crf}
\alias{crf}
\title{Linear-chain Conditional Random Field}
\usage{
crf(x, y, group, method = c("lbfgs", "l2sgd", "averaged-perceptron",
  "passive-aggressive", "arow"), options = crf_options(method)$default,
  file = "annotator.crfsuite", trace = FALSE)
}
\arguments{
\item{x}{a character matrix of data containing attributes about the label sequence \code{y} or an object which can be coerced to a character matrix}

\item{y}{a character vector with the sequence of labels to model}

\item{group}{an integer vector of the same length as \code{y} indicating the group the sequence \code{y} belongs to (e.g. a document or sentence identifier)}

\item{method}{character string with the type of training method. Either one of:
\itemize{
 \item{lbfgs: }{L-BFGS with L1/L2 regularization}
 \item{l2sgd: }{SGD with L2-regularization}
 \item{averaged-perceptron: }{Averaged Perceptron}
 \item{passive-aggressive: }{Passive Aggressive}
 \item{arow: }{Adaptive Regularization of Weights (AROW)}
}}

\item{options}{a list of options to provide to the training algorithm. See \code{\link{crf_options}} for possible options and the example below on how to provide them.}

\item{file}{a character string with the path to the file on disk where the CRF model will be stored.}

\item{trace}{a logical indicating to show the trace of the training output. Defaults to \code{FALSE}.}
}
\value{
an object of class crf which is a list with elements
\itemize{
 \item{method: }{The training method}
 \item{type: }{The type of graphical model which is always set crf1d: Linear-chain (first-order Markov) CRF}
 \item{labels: }{The training labels}
 \item{options: }{A data.frame with the training options provided to the algorithm}
 \item{file_model: }{The path where the CRF model is stored}
 \item{attribute_names: }{The column names of \code{x}}
 \item{log: }{The training log of the algorithm}
}
}
\description{
Fits a Linear-chain (first-order Markov) CRF on the provided label sequence and saves it on disk in order to do sequence labelling.
}
\examples{
library(data.table)
## Get some training data
x <- ner_download_modeldata("conll2002-nl")
x <- as.data.table(x)
x <- x[, pos_previous := shift(pos, n = 1, type = "lag"), by = list(doc_id)]
x <- x[, pos_next := shift(pos, n = 1, type = "lead"), by = list(doc_id)]
x <- as.data.frame(x)
crf_train <- subset(x, data == "ned.train")
crf_test <- subset(x, data == "testa")

## Build the CRF model
opts <- crf_options("lbfgs")
opts <- opts$default
opts$max_iterations <- 5
model <- crf(y = crf_train$label, 
             x = crf_train[, c("token", "pos", "pos_previous", "pos_next")], 
             group = crf_train$doc_id, 
             method = "lbfgs", options = opts, trace = TRUE) 
model
stats <- summary(model, "modeldetails.txt")
stats
plot(stats$iterations$loss)

## Use the CRF model to label a sequence
scores <- predict(model, 
                  newdata = crf_test[, c("token", "pos", "pos_previous", "pos_next")], 
                  group = crf_test$doc_id)
head(scores)
crf_test$label <- scores$label

## cleanup for CRAN
file.remove(model$file_model)
file.remove("modeldetails.txt")
}
\references{
More details about this model is available at \url{http://www.chokkan.org/software/crfsuite}.
}
