% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{crf_cbind_attributes}
\alias{crf_cbind_attributes}
\title{Enrich a data.frame by adding CRF attributes}
\usage{
crf_cbind_attributes(data, field, by, from = -2, to = 2, ngram_max = 3,
  sep = "-")
}
\arguments{
\item{data}{a data.frame which will be coerced to a data.table (cbinding will be done by reference on the existing data.frame)}

\item{field}{a character vector of column names which are part of \code{data} for which we will look for preceding/following terms in the sequence}

\item{by}{a character vector of column names which are part of \code{data} indicating the fields which define the sequence. Preceding/following terms
will be looked for within data of \code{by}.}

\item{from}{integer, by default set to -2, indicating to look up to 2 terms before the current term}

\item{to}{integer, by default  set to 2, indicating to look up to 2 terms after the current term}

\item{ngram_max}{integer indicating the maximum number of terms to combine (2 means bigrams, 3 trigrams, ...)}

\item{sep}{character indicating how to combine the previous/next/current terms}
}
\description{
CRF attributes are merely the neighbouring information of a certain field.
For example the previous word, the next word, the combination of the previous 2 words.
This function cbinds these neighbouring attributes as columns to the provided data.frame.\cr

By default it adds the following column to the data.frame
\itemize{
 \item{the term itself \code{(field[t])}}
 \item{the next term \code{(field[t+1])}}
 \item{the term after that \code{(field[t+2])}}
 \item{the previous term \code{(field[t-1])}}
 \item{the term before the previous term \code{(field[t-2])}}
 \item{as well as all combinations of these terms (bigrams/trigrams/...) where up to \code{ngram_max}
number of terms are combined.}
}
See the example.
}
\examples{
x <- data.frame(doc_id = sort(sample.int(n = 10, size = 1000, replace = TRUE)))
x$pos <- sample(c("Art", "N", "Prep", "V", "Adv", "Adj", "Conj", 
                  "Punc", "Num", "Pron", "Int", "Misc"), 
                  size = nrow(x), replace = TRUE)
x <- crf_cbind_attributes(x, field = "pos", by = "doc_id", 
                          from = -1, to = 1, ngram_max = 3)
head(x)

\dontrun{
## Example on some real data
x <- ner_download_modeldata("conll2002-nl")
x <- crf_cbind_attributes(x, field = c("token", "pos"), 
                          by = c("doc_id", "sentence_id"),
                          ngram_max = 3, sep = "|")
}
}
