% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{crf_entities}
\alias{crf_entities}
\title{Tag a sequence of entities into 1 category}
\usage{
crf_entities(x, entities)
}
\arguments{
\item{x}{a character vector of categories in the sequence of occurring (e.g. B-LOC, I-LOC, I-PER, B-PER, O, O, B-PER)}

\item{entities}{a list of groups, where each list element contains
\itemize{
 \item{start: }{A length 1 character string with the start element identifying a sequence. E.g. B-LOC}
 \item{labels: }{A character vector containing all the elements which are considered being part of a same labelling sequence, including the starting element. 
 E.g. \code{c('B-LOC', 'I-LOC', 'E-LOC')}}
}
The list name will define a label to the entity}
}
\value{
as list with elements \code{id_category} and \code{category} where 
\itemize{
 \item{category is a character vector of entities, constructed by recoding \code{x} to the names of \code{names(entities})}
 \item{id_category is an integer vector containing unique identifiers identfying the compound label sequence such that e.g. the sequence 'B-LOC', 'I-LOC', 'I-LOC', 'E-LOC'
 (New York City District) would get the same \code{id_category} identifier.}
}
See the examples.
}
\description{
Named Entity Recognition or Chunking labelling following the beginning/continues/end scheme are combined into 1
entity. An example would be 'New', 'York', 'City', 'District' which can be labbelled as 'B-LOC', 'I-LOC', 'I-LOC', 'E-LOC'.
The function looks for this sequence starting with 'B-LOC' and combines all subsequent labels into 1 category and assigns a unique identifier.
}
\examples{
library(data.table)
x <- ner_download_modeldata("conll2002-nl")
x <- as.data.table(x)

##
## Define entity groups 
## and combine B-LOC I-LOC I-LOC sequences as 1 group (e.g. New York City) 
groups <- list(
 Location = list(start = "B-LOC", labels = c("B-LOC", "I-LOC", "E-LOC")),
 Organisation =  list(start = "B-ORG", labels = c("B-ORG", "I-ORG", "E-ORG")),
 Person = list(start = "B-PER", labels = c("B-PER", "I-PER", "E-PER")), 
 Misc = list(start = "B-MISC", labels = c("B-MISC", "I-MISC", "E-MISC")))
x[, c("id", "entity") := crf_entities(label, entities = groups), by = "doc_id"]

## Show some organisations
subset(x, entity == "Organisation")
}
