% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{predict.crf}
\alias{predict.crf}
\title{Predict the label sequence based on the Conditional Random Field}
\usage{
\method{predict}{crf}(object, newdata, group, type = c("marginal",
  "sequence"), trace = FALSE, ...)
}
\arguments{
\item{object}{an object of class crf as returned by \code{\link{crf}}}

\item{newdata}{a character matrix of data containing attributes about the label sequence \code{y} or an object which can be coerced to a character matrix. 
This data should be provided in the same format as was used for training the model}

\item{group}{an integer vector of the same length as nrow \code{newdata} indicating the group the sequence \code{y} belongs to (e.g. a document or sentence identifier)}

\item{type}{either 'marginal' or 'sequence' to get predictions at the level of \code{newdata} or a the level of the sequence \code{group}. Defaults to \code{'marginal'}}

\item{trace}{a logical indicating to show the trace of the labelling output. Defaults to \code{FALSE}.}

\item{...}{not used}
}
\value{
If \code{type} is 'marginal': a data.frame with columns label and marginal containing the viterbi decoded predicted label and marginal probability. \cr
If \code{type} is 'sequence': a data.frame with columns group and probability containing for each sequence group the probability of the sequence.
}
\description{
Predict the label sequence based on the Conditional Random Field
}
\seealso{
\code{\link{crf}}
}
