% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{sprintf_na}
\alias{sprintf_na}
\title{NA friendly version of sprintf}
\usage{
sprintf_na(fmt, ...)
}
\arguments{
\item{fmt}{a character vector of format strings, which will be fed on to \code{\link[base]{sprintf}}}

\item{...}{values to be passed into \code{fmt}, which will be passed on to \code{\link[base]{sprintf}}}
}
\value{
A character vector of length that of the longest input. Same as what \code{\link[base]{sprintf}}.
In case any of the values passed on to \code{...}, the corresponding returned value will be set to NA.
}
\description{
Does the same as the function \code{\link[base]{sprintf}} except that if 
in ... NA values are passed, also NA values are returned.
}
\examples{
sprintf("(w-1):\%s", c("xyz", NA, "abc"))
sprintf_na("(w-1):\%s", c("xyz", NA, "abc"))
sprintf("(w-1):\%s_\%s", c("xyz", NA, "abc"), c(NA, "xyz", "abc"))
sprintf_na("(w-1):\%s_\%s", c("xyz", NA, "abc"), c(NA, "xyz", "abc"))
}
