% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{crf_evaluation}
\alias{crf_evaluation}
\title{Basic classification evaluation metrics for multi-class labelling}
\usage{
crf_evaluation(pred, obs, labels = na.exclude(unique(c(as.character(pred),
  as.character(obs)))), labels_overall = setdiff(labels, "O"))
}
\arguments{
\item{pred}{a factor with predictions}

\item{obs}{a factor with gold labels}

\item{labels}{a character vector of possible values that \code{pred} and \code{obs} can take. Defaults to the values in the data}

\item{labels_overall}{a character vector of either labels which is either the same as \code{labels} or a subset of \code{labels} in order to compute a weighted average of the by-label statistics}
}
\value{
a list with 2 elements:
\enumerate{
\item{bylabel: data.frame with the precision, recall, specificity, F1 score and support for each label}
\item{overall: a vector with the overall accuracy and the precision, recall, specificity and F1 score weighted by the support 
for all the labels provided in \code{labels_overall} as well as just the average of these 4 metrics giving equal weight to each label}
}
}
\description{
The precision, recall, specificity, F1 measure and support metrics are provided for each label in a one-versus rest setting.
}
\examples{
pred <- sample(LETTERS, 1000, replace = TRUE)
gold <- sample(LETTERS, 1000, replace = TRUE)
crf_evaluation(pred = pred, obs = gold, labels = LETTERS) 

\dontrun{
x <- ner_download_modeldata("conll2002-nl")
x <- crf_cbind_attributes(x, terms = c("token", "pos"), 
                          by = c("doc_id", "sentence_id"))
crf_train <- subset(x, data == "ned.train")
crf_test <- subset(x, data == "testa")
attributes <- grep("token|pos", colnames(x), value=TRUE)
model <- crf(y = crf_train$label, 
             x = crf_train[, attributes], 
             group = crf_train$doc_id, 
             method = "lbfgs") 
             
## Use the model to score on existing tokenised data
scores <- predict(model, 
                  newdata = crf_test[, attributes], 
                  group = crf_test$doc_id)
crf_evaluation(pred = scores$label, obs = crf_test$label)
crf_evaluation(pred = scores$label, obs = crf_test$label, 
  labels = c("O", 
             "B-ORG", "I-ORG", "B-PER", "I-PER", 
             "B-LOC", "I-LOC", "B-MISC", "I-MISC"))
             
             
library(udpipe)
pred <- txt_recode(scores$label, 
                   from = c("B-ORG", "I-ORG", "B-PER", "I-PER", 
                            "B-LOC", "I-LOC", "B-MISC", "I-MISC"),
                   to = c("ORG", "ORG", "PER", "PER", 
                          "LOC", "LOC", "MISC", "MISC"))
obs <- txt_recode(crf_test$label, 
                  from = c("B-ORG", "I-ORG", "B-PER", "I-PER", 
                           "B-LOC", "I-LOC", "B-MISC", "I-MISC"),
                  to = c("ORG", "ORG", "PER", "PER", 
                         "LOC", "LOC", "MISC", "MISC"))
crf_evaluation(pred = pred, obs = obs, 
               labels = c("ORG", "LOC", "PER", "MISC", "O"))
}
}
